Fix pointer arithmetic

+++ src/bitmap.c
@@ -24,7 +24,9 @@
 */
 
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 #include <noftypes.h>
 #include <bitmap.h>
 
@@ -61,7 +63,7 @@ static bitmap_t *_make_bitmap(uint8 *data_addr, bool hw, int width,
    if (false == bitmap->hardware)
    {
       bitmap->pitch = (bitmap->pitch + 3) & ~3;
-      bitmap->line[0] = (uint8 *) (((uint32) bitmap->data + overdraw + 3) & ~3);
+      bitmap->line[0] = (uint8 *) (((uintptr_t) bitmap->data + overdraw + 3) & ~3);
    }
    else
    {
